Selection sort

+++ 2-selection_sort.c
@@ -11,7 +11,7 @@ int current;
 size_t i, j;
 for (i = 0; i < size; i++)
 {
-int min = i;
+size_t min = i;
 for (j = i + 1; j < size; j++)
 {
 if (array[j] < array[min])
